Version 20240219.1
https://www.gurux.fi/release/guruxdlmscpp-202402191

+++ development/src/GXBigInteger.cpp
@@ -756,7 +756,7 @@ std::string CGXBigInteger::ToString()
         {
             bb.SetUInt32(Data[pos]);
         }
-        for (pos = 0; pos != bb.GetSize(); ++pos)
+        for (pos = 0; pos != (int) bb.GetSize(); ++pos)
         {
             if (bb.GetData()[pos] != 0)
             {
